Add files via upload
The data analysis script contains all the statistical analyses and allows for the reconstruction of the figures presented in the article. Additionally, the database is accessible. However, within the Python script, there is a specific cell designed to filter the appropriate population used in the analysis, as the experiment was conducted online, resulting in numerous incomplete sessions that need to be excluded.